% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\name{create_dictionary}
\alias{create_dictionary}
\title{create_dictionary}
\usage{
create_dictionary(
  in_file,
  out_file = "",
  data_ext = ".csv",
  output_ext = "_dict.csv"
)
}
\arguments{
\item{in_file}{data file to describe}

\item{out_file}{dictionary file to create, uses defaults to create}

\item{data_ext}{optional: extension for data to use when creating output file name (.csv)}

\item{output_ext}{optional: replacement extension (_dict.csv)}
}
\description{
create a data dictionary csv file based on data input
will not update an existing one
}
\examples{
create_dictionary(in_file = "my_data.csv")
}
