% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\name{dictionaries_to_singlepage}
\alias{dictionaries_to_singlepage}
\title{dictionaries_to_singlepage}
\usage{
dictionaries_to_singlepage(
  index_file = "_index.md",
  in_folder = "./input",
  start_tpl = "<!-- START {f} -->",
  end_tpl = "<!-- END {f} -->"
)
}
\arguments{
\item{index_file}{file to use as new or replacement, defaults to '_index.md'}

\item{in_folder}{folder to search, defaults to 'input'}

\item{start_tpl}{starting block comment template, defaults to '<!-- START {f} -->'}

\item{end_tpl}{ending block comment template, defaults to '<!-- END {f} -->'}
}
\description{
combine dictionary markdown files in a folder into a single updateable index
with an existing file, it will replace replace comment-marked text blocks
without changing the rest of the document
}
\details{
designed to work with output from dictionary_to_markdown()

also designed not to interfere with markdown previews as comments are hidden
}
\examples{
dictionaries_to_singlepage(in_file = "readme.md", in_folder = "./input")
}
