% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columndropper.R
\name{dropcolumns}
\alias{dropcolumns}
\title{dropcolumns
Function to filter the columns in the dataframe}
\usage{
dropcolumns(dfname, colslist)
}
\arguments{
\item{dfname}{Dataframe}

\item{colslist}{List of column names to drop}
}
\value{
Dataframe without those columns
}
\description{
dropcolumns
Function to filter the columns in the dataframe
}
\examples{
temp1 <- dropcolumns(df, c("One", "Two", "Three"))
}
