% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filereader.R
\name{fast_read_csv}
\alias{fast_read_csv}
\title{fast_read_csv
Function to open a file via data.table - faster}
\usage{
fast_read_csv(fname, data.table = FALSE, ...)
}
\arguments{
\item{fname}{The filename as a string}

\item{data.table}{Default (FALSE) Use data.table=TRUE to return a data.table}

\item{...}{Additional parameters supported by data.table::fread,}
}
\value{
Data.table for the file
}
\description{
fast_read_csv
Function to open a file via data.table - faster
}
\examples{
temp1 <- fast_read_csv("../../Files/thing.csv")
temp2 <- fast_read_csv(paste0("datapath", "filename"))
}
