% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filereader.R
\name{filereader}
\alias{filereader}
\title{filereader
Function to identify whether an input file is of type
csv, dbf or xlsx, and open it accordingly}
\usage{
filereader(fname)
}
\arguments{
\item{fname}{The filename as a string}
}
\value{
Dataframe for the file
}
\description{
filereader
Function to identify whether an input file is of type
csv, dbf or xlsx, and open it accordingly
}
