% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columndropper.R
\name{keepcolumns}
\alias{keepcolumns}
\title{keepcolumns
Function to filter the columns in the dataframe}
\usage{
keepcolumns(dfname, colslist)
}
\arguments{
\item{dfname}{Dataframe}

\item{colslist}{List of column names to keep}
}
\value{
Dataframe with only those columns
}
\description{
keepcolumns
Function to filter the columns in the dataframe
}
\examples{
temp1 <- keepcolumns(df, c("One", "Two", "Three"))
}
