% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadPackages}
\alias{loadPackages}
\title{Load package, install package if not installed}
\usage{
loadPackages(package_list)
}
\arguments{
\item{package_list}{list of packages}
}
\description{
Load package, install package if not installed
}
