% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{pkg_rd2markdown}
\alias{pkg_rd2markdown}
\title{pkg_rd2markdown}
\usage{
pkg_rd2markdown(in_folder = "man", out_folder = "docs", ...)
}
\arguments{
\item{in_folder}{path to man/ where .Rd files created by devtools::documentation/oxygen2
defaults to in_folder="man", out_folder="docs"}
}
\description{
convenience function to wrap around rd2markdown::rd2markdown()
walks a folder for all Rd files and converts to
converts all Rd files in a folder to regular markdown for easier viewing in gitlab/github
outputs a list of links that can be pasted into a readme
}
\examples{
pkg_rd2markdown("man")
}
