% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_withpath.R
\name{write_csv_withpath}
\alias{write_csv_withpath}
\title{write_csv_withpath
Function to write files and also create folder if it doesn't yet exist.}
\usage{
write_csv_withpath(df, fname, ...)
}
\arguments{
\item{df}{Dataframe to write to file}

\item{fname}{The filename as a string}

\item{...}{Additional parameters supported by data.table::fwrite}
}
\description{
write_csv_withpath
Function to write files and also create folder if it doesn't yet exist.
}
\examples{
temp1 <- write_csv_withpath(df, "../../Files/thing.csv")
}
