% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\name{dictionary_to_markdown}
\alias{dictionary_to_markdown}
\title{dictionary_to_markdown
convert a dictionary file to a markdown document
for pasting into documentation
csv only}
\usage{
dictionary_to_markdown(dictionary_file, in_file = "")
}
\arguments{
\item{dictionary_file}{path to dictionary file}

\item{in_file}{(optional) path to original data file}
}
\description{
columns expected: COLUMN,TYPE,DESCRIPTION
}
\examples{
dictionary_to_markdown("../../my_data_dict.csv")
}
